backend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Database
psycopg2-binary==2.9.7

# Validation
marshmallow==3.20.1
email-validator==2.0.0

# Security
bcrypt==4.0.1
cryptography==41.0.4

# Environment management
python-dotenv==1.0.0

# Logging and monitoring
python-json-logger==2.0.7

# Development utilities
Werkzeug==2.3.7
